refactor: Extract shared sans-I/O cores from sync data paths
